refactor: extract report struct into models
